InvChisquare {geoR}	R Documentation

The (Scaled) Inverse Chi-Squared Distribution

Description

Density and random generation for the scaled inverse chi-squared (chi^2_{ScI}) distribution with df degrees of freedom and optional non-centrality parameter scale.

Usage

dinvchisq(x, df, scale, log = FALSE)
rinvchisq(n, df, scale = 1/df)

Arguments

x
vector of quantiles.

n
number of observations. If length(n) > 1, the length is taken to be the number required.

df
degrees of freedom.

scale
scale parameter.

log
logical; if TRUE, densities d are given as log(d).

Details

The inverse chi-squared distribution with df= n degrees of freedom has density

f(x) = 1 / (2^(n/2) Gamma(n/2)) (1/x)^(n/2-1) e^(-1/(2x))

for x > 0. The mean and variance are 1/(n-2) and 2/((n-4)(n-2)^2).

The non-central chi-squared distribution with df= n degrees of freedom and non-centrality parameter scale = S^2 has density

f(x) = ((n/2)^(n/2))/(Γ (n/2)) S^n (1/x)^((n/2)+1) e^(-(n S^2)/(2x))

, for x ≥ 0. The first is a particular case of the latter for λ = n/2.

Value

dinvchisq gives the density and rinvchisq generates random deviates.

See Also

rchisq for the chi-squared distribution which is the basis for this function.

Examples

set.seed(1234); rinvchisq(5, df=2)
set.seed(1234); 1/rchisq(5, df=2)

set.seed(1234); rinvchisq(5, df=2, scale=5)
set.seed(1234); 5*2/rchisq(5, df=2)

## inverse Chi-squared is a particular case
x <- 1:10
all.equal(dinvchisq(x, df=2), dinvchisq(x, df=2, scale=1/2))

[Package geoR version 1.7-5.2.1 Index]