Ksat {geoR}	R Documentation

Saturated Hydraulic Conductivity

Description

The data consists of 32 measurements of the saturated hydraulic conductivity of a soil.

Usage

data(Ksat)

Format

The object Ksat is a list of the class geodata with the following elements:

coords

a matrix with the coordinates of the soil samples.

data

measurements of the saturated hidraulic conductivity.

borders

a data-frame with the coordinates of a polygon defining the borders of the area.

Source

Data provided by Dr. Décio Cruciani, ESALQ/USP, Brasil.

Examples

summary(Ksat)
plot(Ksat, border=borders)

[Package geoR version 1.7-5.2.1 Index]