boxcox.geodata {geoR}	R Documentation

Box-Cox transformation for geodata objects

Description

Method for Box-Cox transformation for objects of the class geodata assuming the data are independent. Computes and optionally plots profile log-likelihoods for the parameter of the Box-Cox simple power transformation y^lambda.

Usage

## S3 method for class 'geodata'
boxcox(object, trend = "cte", ...)

Arguments

object
an object of the class geodata. See as.geodata.

trend
specifies the mean part of the model. See trend.spatial for further details. Defaults to "cte".

...
arguments to be passed for the function boxcox.

Details

This is just a wrapper for the function boxcox facilitating its usage with geodata objects.

Notice this assume independent observations which is typically not the case for geodata objects.

Value

A list of the lambda vector and the computed profile log-likelihood vector, invisibly if the result is plotted.

See Also

boxcox for parameter estimation results for independent data and likfit for parameter estimation within the geostatistical model.

Examples

if(require(MASS)){
boxcox(wolfcamp)

data(ca20)
boxcox(ca20, trend = ~altitude)
}

[Package geoR version 1.7-5.2.1 Index]