boxcoxfit {geoR}	R Documentation

Parameter Estimation for the Box-Cox Transformation

Description

Parameter estimation and plotting of the results for the Box-Cox transformed normal distribution.

Usage

boxcoxfit(object, xmat, lambda, lambda2 = NULL, add.to.data = 0, ...)

## S3 method for class 'boxcoxfit'
print(x, ...)

## S3 method for class 'boxcoxfit'
plot(x, hist = TRUE, data = eval(x$call$object), ...)

## S3 method for class 'boxcoxfit'
lines(x, data = eval(x$call$object), ...)

Arguments

object
a vector with the data.

xmat
a matrix with covariates values. Defaults to rep(1, length(y)).

lambda
numerical value(s) for the transformation parameter lambda. Used as the initial value in the function for parameter estimation. If not provided default values are assumed. If multiple values are passed the one with highest likelihood is used as initial value.

lambda2
logical or numerical value(s) of the additional transformation (see DETAILS below). Defaults to NULL. If TRUE this parameter is also estimated and the initial value is set to the absolute value of the minimum data. A numerical value is provided it is used as the initial value. Multiple values are allowed as for lambda.

add.to.data
a constant value to be added to the data.

x
a list, typically an output of the function boxcoxfit.

hist
logical indicating whether histograms should to be plotted.

data
data values.

...
extra parameters to be passed to the minimization function optim (boxcoxfit), hist (plot) or curve (lines).

Value

The functions returns the following results:

boxcoxfit
a list with estimated parameters and results on the numerical minimization.

print.boxcoxfit
print estimated parameters. No values returned.

plot.boxcoxfit
plots histogram of the data (optional) and the model. No values returned. This function is only valid if covariates are not included in boxcoxfit.

lines.boxcoxfit
adds a line with the fitted model to the current plot. No values returned. This function is only valid if covariates are not included in boxcoxfit.

Author(s)

Paulo Justiniano Ribeiro Jr. paulojus@leg.ufpr.br,
Peter J. Diggle p.diggle@lancaster.ac.uk.

References

Box, G.E.P. and Cox, D.R.(1964) An analysis of transformations. JRSS B 26:211–246.

See Also

rboxcox and dboxcox for the expression and more on the Box-Cox transformation, the minimization function optim, the function boxcox in the package MASS and the function box.cox in the package car.

Examples

set.seed(384)
## Simulating data
simul <- rboxcox(100, lambda=0.5, mean=10, sd=2)
## Finding the ML estimates
ml <- boxcoxfit(simul)
ml
## Ploting histogram and fitted model
plot(ml)
##
## Comparing models with different lambdas,
## zero  means and unit variances
curve(dboxcox(x, lambda=-1), 0, 8)
for(lambda in seq(-.5, 1.5, by=0.5))
  curve(dboxcox(x, lambda), 0, 8, add = TRUE)
##
## Another example, now estimating lambda2
##
simul <- rboxcox(100, lambda=0.5, mean=10, sd=2)
ml <- boxcoxfit(simul, lambda2 = TRUE)
ml
plot(ml)
##
## An example with a regression model
##
boxcoxfit(object = trees[,3], xmat = trees[,1:2])

[Package geoR version 1.7-5.2.1 Index]