elevation {geoR}	R Documentation

Surface Elevations

Description

Surface elevation data taken from Davis (1972). An onject of the class geodata with elevation values at 52 locations.

Usage

data(elevation)

Format

An object of the class geodata which is a list with the following elements:

coords

x-y coordinates (multiples of 50 feet).

data

elevations (multiples of 10 feet).

Source

Davis, J.C. (1973) Statistics and Data Analysis in Geology. Wiley.

Examples

summary(elevation)
plot(elevation)

[Package geoR version 1.7-4 Index]