eyefit {geoR}	R Documentation

Interactive Variogram Estimation

Description

Function to fit an empirical variogram "by eye" using an interactive Tcl-Tk interface.

Usage

eyefit(vario, silent = FALSE)

Arguments

vario
An empirical variogram object as returned by the function variog.

silent
logical indicating wheather or not the fitted variogram must be returned.

Value

Returns a list of list with the model parameters for each of the saved fit(s).

Author(s)

Andreas Kiefer andreas@inf.ufpr.br
Paulo Justiniano Ribeiro Junior paulojus@leg.ufpr.br.

See Also

variofit for least squares variogram fit, likfit for likelihood based parameter estimation and krige.bayes to obtain the posterior distribution for the model parameters.

[Package geoR version 1.7-5.2.1 Index]