gambia {geoR}	R Documentation

Gambia Malaria Data

Description

Malaria prevalence in children recorded at villages in The Gambia, Africa.

Usage

data(gambia)

Format

Two objects are made available:

gambia
A data frame with 2035 observations on the following 8 variables.

x

x-coordinate of the village (UTM).

y

y-coordinate of the village (UTM).

pos

presence (1) or absence (0) of malaria in a blood sample taken from the child.

age

age of the child, in days

netuse

indicator variable denoting whether (1) or not (0) the child regularly sleeps under a bed-net.

treated

indicator variable denoting whether (1) or not (0) the bed-net is treated (coded 0 if netuse=0).

green

satellite-derived measure of the green-ness of vegetation in the immediate vicinity of the village (arbitrary units).

phc

indicator variable denoting the presence (1) or absence (0) of a health center in the village.

gambia.borders
A data frame with 2 variables:

x

x-coordinate of the country borders.

y

y-coordinate of the country borders.

References

Thomson, M., Connor, S., D Alessandro, U., Rowlingson, B., Diggle, P., Cresswell, M. & Greenwood, B. (1999). Predicting malaria infection in Gambian children from satellite data and bednet use surveys: the importance of spatial correlation in the interpretation of results. American Journal of Tropical Medicine and Hygiene 61: 2–8.

Diggle, P., Moyeed, R., Rowlingson, B. & Thomson, M. (2002). Childhood malaria in The Gambia: a case-study in model-based geostatistics, Applied Statistics.

Examples

plot(gambia.borders, type="l", asp=1)
points(gambia[,1:2], pch=19)
# a built-in function for a zoomed map
gambia.map()
# Building a "village-level" data frame
ind <- paste("x",gambia[,1], "y", gambia[,2], sep="")
village <- gambia[!duplicated(ind),c(1:2,7:8)]
village$prev <- as.vector(tapply(gambia$pos, ind, mean))
plot(village$green, village$prev)

[Package geoR version 1.7-4 Index]