hist.krige.bayes {geoR}	R Documentation

Plots Sample from Posterior Distributions

Description

Plots histograms and/or density estimation with samples from the posterior distribution of the model parameters

Usage

## S3 method for class 'krige.bayes'
hist(x, pars, density.est = TRUE, histogram = TRUE, ...)

Arguments

x
an object of the class krige.bayes, with an output of the funtions krige.bayes.

pars
a vector with the names of one or more of the model parameters. Defaults to all model parameters. Setting to -1 excludes the intercept.

density.est
logical indication whether a line with the density estimation should be added to the plot.

histogram
logical indicating whether the histogram is included in the plot.

...
further arguments for the plotting functions and or for the density estimation.

Value

Produces a plot in the currently graphics device.
Returns a invisible list with the components:

histogram
with the output of the function hist for each parameter

density.estimation
with the output of the function density for each parameter

Author(s)

Paulo J. Ribeiro Jr. paulojus@leg.ufpr.br,
Peter J. Diggle p.diggle@lancaster.ac.uk.

See Also

krige.bayes, hist, density.

Examples

## See documentation for krige.bayes()

[Package geoR version 1.7-5.2.1 Index]