hoef {geoR}	R Documentation

Data for spatial analysis of experiments

Description

The hoef data frame has 25 rows and 5 columns.
The data consists of a uniformity trial for which artificial treatment effects were assign to the plots.

Usage

data(hoef)

Format

This data frame contains the following columns:

x1

x-coordinate of the plot.

x2

y-coordinate of the plot.

dat

the artificial data.

trat

the treatment number.

ut

the data from the uniformity trial, without the treatment effect.

Details

The treatment effects assign to the plots are:

Treatment 1: tau1 = 0

Treatment 2: tau2 = -3

Treatment 3: tau3 = -5

Treatment 4: tau4 = 6

Treatment 5: tau5 = 6

Source

Ver Hoef, J.M. & Cressie, N. (1993) Spatial statistics: analysis of field experiments. In Scheiner S.M. and Gurevitch, J. (Eds) Design and Analysis of Ecological Experiments. Chapman and Hall.

Examples

hoef.geo <- as.geodata(hoef, covar.col=4)
summary(hoef)
summary(hoef.geo)
points(hoef.geo, cex.min=2, cex.max=2, pt.div="quintiles")

[Package geoR version 1.7-4 Index]