image.grf {geoR}	R Documentation

Image, Contour or Perspective Plot of Simulated Gaussian Random Field

Description

Methods for image, contour or perspective plot of a realisation of a Gaussian random field, simulated using the function grf.

Usage

## S3 method for class 'grf'
image(x, sim.number = 1, borders, x.leg, y.leg, ...)
## S3 method for class 'grf'
contour(x, sim.number = 1, borders, filled = FALSE, ...)
## S3 method for class 'grf'
persp(x, sim.number = 1, borders, ...)

Arguments

x
an object of the class grf, typically an output of the function grf.

sim.number
simulation number. Indicates the number of the simulation top be plotted. Only valid if the object contains more than one simulation. Defaults to 1.

borders
optional. Typically a two coluns matrix especifying a polygon. Points outside the borders will be set no NA

x.leg, y.leg
limits for the legend in the horizontal and vertical directions.

filled
logical. If FALSE the function contour is used otherwise filled.contour. Defaults to FALSE.

...
further arguments to be passed to the functions image, contour or persp.

Value

An image or perspective plot is produced on the current graphics device. No values are returned.

Author(s)

Paulo Justiniano Ribeiro Jr. paulojus@leg.ufpr.br,
Peter J. Diggle p.diggle@lancaster.ac.uk.

References

Further information about the package geoR can be found at:
http://www.leg.ufpr.br/geoR.

See Also

grf for simulation of Gaussian random fields, image and persp for the generic plotting functions.

Examples

# generating 4 simulations of a Gaussian random field
sim <- grf(441, grid="reg", cov.pars=c(1, .25), nsim=4)
op <- par(no.readonly = TRUE)
par(mfrow=c(2,2), mar=c(3,3,1,1), mgp = c(2,1,0))
for (i in 1:4)
  image(sim, sim.n=i)
par(op)

[Package geoR version 1.7-5.2.1 Index]