image.krige.bayes {geoR}	R Documentation

Plots Results of the Predictive Distribution

Description

This function produces an image or perspective plot of a selected element of the predictive distribution returned by the function krige.bayes.

Usage

## S3 method for class 'krige.bayes'
image(x, locations, borders,
                  values.to.plot=c("mean", "variance",
                            "mean.simulations", "variance.simulations",
                            "quantiles", "probabilities", "simulation"),
                  number.col, coords.data, x.leg, y.leg, messages, ...)
## S3 method for class 'krige.bayes'
contour(x, locations, borders, 
                  values.to.plot = c("mean", "variance",
                       "mean.simulations", "variance.simulations",
                       "quantiles", "probabilities", "simulation"),
                  filled=FALSE, number.col, coords.data,
                  x.leg, y.leg, messages, ...)
## S3 method for class 'krige.bayes'
persp(x, locations, borders,
                  values.to.plot=c("mean", "variance",
                       "mean.simulations", "variance.simulations",
                       "quantiles", "probabilities", "simulation"),
                  number.col, messages, ...)

Arguments

x
an object of the class krige.bayes, typically an output of the function krige.bayes.

locations
an n x 2 matrix with the coordinates of the prediction locations, which should define a regular grid in order to be plotted by image or persp. By default does not need to be provided and evaluates the attribute "prediction.locations" from the input object.

borders
an n x 2 matrix with the coordinates defining the borders of a region inside the grid defined by locations. Elements in the argument values are assigned to locations internal to the borders and NA's to the external ones.

values.to.plot
select the element of the predictive distribution to be plotted. See DETAILS below.

filled
logical. If FALSE the function contour is used otherwise filled.contour. Defaults to FALSE.

number.col
Specifies the number of the column to be plotted. Only used if previous argument is set to one of "quantiles", "probabilities" or "simulation".

coords.data
optional. If an n x 2 matrix with the data coordinates is provided, points indicating the data locations are included in the plot.

x.leg, y.leg
limits for the legend in the horizontal and vertical directions.

messages
logical, if TRUE status messages are printed while running the function.

...
extra arguments to be passed to the plotting function image or persp.

Details

The function krige.bayes returns summaries and other results about the predictive distributions. The argument values.to.plot specifies which result will be plotted. It can be passed to the function in two different forms:

a vector with the object containing the values to be plotted, or

one of the following options: "moments.mean", "moments.variance", "mean.simulations", "variance.simulations", "quantiles", "probability" or "simulation".

For the last three options, if the results are stored in matrices, a column number must be provided using the argument number.col.

The documentation for the function krige.bayes provides further details about these options.

Value

An image or persp plot is produced on the current graphics device. No values are returned.

Author(s)

Paulo J. Ribeiro Jr. paulojus@leg.ufpr.br,
Peter J. Diggle p.diggle@lancaster.ac.uk.

References

Further information on the package geoR can be found at:
http://www.leg.ufpr.br/geoR.

See Also

krige.bayes for Bayesian Kriging computations and, image and persp for the generic plotting functions.

Examples

#See examples in the documentation for the function krige.bayes().

[Package geoR version 1.7-5.2.1 Index]