kattegat {geoR}	R Documentation

Kattegat basin salinity data

Description

Salinity measurements at the Kattegat basin, Denmark.

Usage

data(kattegat)

Format

An object of the class "geodata", which is list with three components:

coords

the coordinates of the data locations. The distance are given in kilometers.

data

values of the piezometric head. The unit is heads to meters.

dk

a list with cooordinates of lines defining borders and islands across the study area.

Source

National Environmental Research Institute, Arhus University, Denmark and the Swedish Meteorological and Hydrological Institute.

References

Diggle, P. J. and Lophaven, S. (2006). Bayesian geostatistical design. Scandinavian Journal of Statistics, 33: 55-64.

Examples

plot(c(550,770),c(6150,6420),type="n",xlab="X UTM",ylab="Y UTM")
points(kattegat, add=TRUE)
lapply(kattegat$dk, lines, lwd=2)

[Package geoR version 1.7-4 Index]