landim1 {geoR}	R Documentation

Data from Landim's book

Description

Artificial or non-specified data from Paulo Landim's book

Usage

data(landim1)

Format

A data frame with 38 observations on the following 4 variables.

EW

a numeric vector with the east-west coordinates.

NS

a numeric vector with the north-south coordinates.

A

a numeric vector with data on a first variable.

B

a numeric vector with data on a second variable.

Source

Landim, P. M. B. (2004) Análise estatística de dados geológicos. Editora Unesp. Data from Table~1, pg.12.

Examples

data(landim)
plot(as.geodata(landim1, data.col=3))
plot(as.geodata(landim1, data.col=4))

[Package geoR version 1.7-5.2.1 Index]