legend.krige {geoR}	R Documentation

Add a legend to a image with kriging results

Description

This function allows adds a legend to an image plot generated by image.kriging or image.krige.bayes. It can be called internally by these functions or directly by the user.

Usage

legend.krige(x.leg, y.leg, values, scale.vals,
             vertical = FALSE, offset.leg = 1, ...)

Arguments

x.leg
limits for the legend in the x direction.

y.leg
limits for the legend in the y direction.

values
values plotted in the image.

scale.vals
optional. Values to appear in the legend. If not provided the function pretty is used to define the values.

vertical
If TRUE the legend is drawn in the vertical direction. Defaults to FALSE.

offset.leg
numeric value controlling the distance between the legend text and the legend box.

...
further arguments to be passed to the function text.

Value

A legend is added to the current plot. No values are returned.

Author(s)

Paulo J. Ribeiro Jr. paulojus@leg.ufpr.br,
Peter J. Diggle p.diggle@lancaster.ac.uk.

References

Further information on the package geoR can be found at:
http://www.leg.ufpr.br/geoR.

See Also

image.kriging, image.krige.bayes.

Examples

# See examples in the documentation for image.kriging

[Package geoR version 1.7-5.2.1 Index]