likfitBGCCM {geoR}	R Documentation

Fits the bivariate Gaussian common component geostatistical model

Description

Computes maximum likelihood estimates of the bivariate Gaussian common component geostatistical model.

Usage

likfitBGCCM(geodata1, geodata2, ini.sigmasq, ini.phi,
            cov0.model="matern", cov1.model="matern", cov2.model="matern",
            kappa0=0.5, kappa1=0.5, kappa2=0.5,
            fc.min = c("optim", "nlminb"), ...)

Arguments

geodata1
an object of the class geodata with the first variable.

geodata2
an object of the class geodata with the second variable.

ini.sigmasq
optional, a vector with initial values for the variance parameters. If not provided default values are used.

ini.phi
optional, a vector with initial values for the correlation range parameters. If not provided default values are used.

cov0.model, cov1.model, cov2.model
covariance model for each of the processes. See cov.spatial for details.

kappa0, kappa1, kappa2
extra parameter for some covariance models.

fc.min
a string indication which function should be used to minimise the negative of the log-likelihood.

...
further arguments to be passed to optim or nlminb.

Value

A list with model fitting information to which the class BGCCM is assigned.

mu
a 2 elements vector with mean estimates.

sigmasq
a 4 elements vector with variance estimates.

phi
a 3 elements vector with estimated correlation parameters values.

loglik
a scalar. Maximised value of the log-likelihood.

optim
results returned by optim or nlminb.

...
and other information related to the model fitting.

Warning

This is a new function and still in draft format and pretty much untested.

Author(s)

Paulo J. Ribeiro Jr. paulojus@leg.ufpr.br,
Peter J. Diggle p.diggle@lancaster.ac.uk.

See Also

optim, nlminb, varcovBGCCM, as.geodata, likfit.

Examples

# see http://www.leg.ufpr.br/geoR/tutorials/CCM.R

[Package geoR version 1.7-5.2.1 Index]