lines.variogram {geoR}	R Documentation

Line with a Empirical Variogram

Description

A sample (empirical) variogram computed using the function variog is added to the current plot.

Usage

## S3 method for class 'variogram'
lines(x, max.dist, type = "o",  scaled = FALSE,
         pts.range.cex, ...)

Arguments

x
an object of the class "variogram", typically an output from the function variog.

max.dist
maximum distance for the x-axis. By default takes the maximum distance for which the sample variogram was computed.

type
type of line for the empirical variogram. The default is "o" (dots and lines). See documentation for lines for further details.

scaled
logical. If TRUE the variogram values are divided by the sample variance. This allows comparison between variograms of different variables.

pts.range.cex
optional. A two elements vector with maximum and minimum values for the caracter expansion factor cex. If provided the point sizes in binned variogram are proportional to the number of pairs of points used to compute each bin.

...
other arguments to be passed to the function lines.

Value

A line with the empirical variogram is added to the plot in the current graphics device. No values are returned.

Author(s)

Paulo Justiniano Ribeiro Jr. paulojus@leg.ufpr.br,
Peter J. Diggle p.diggle@lancaster.ac.uk.

References

Further information on the package geoR can be found at:
http://www.leg.ufpr.br/geoR.

See Also

variog, lines.variogram, lines.variomodel, variog.model.env, variog.mc.env, plot.grf, lines.

[Package geoR version 1.7-5.2.1 Index]