lines.variomodel.grf {geoR}	R Documentation

Lines with True Variogram for Simulated Data

Description

This functions adds to the graphics device a line with the theoretical (true) variogram used when generating simulations with the function grf.

Usage

## S3 method for class 'grf'
lines.variomodel(x, max.dist, n = 100, ...)

Arguments

x
an object from the class grf typically an output of the function grf.

max.dist
maximum distance to compute and plot the true variogram. Defaults to the maximum distance between two data locations.

n
number of points used to compute and draw the variogram line.

...
further arguments to be passed to the function curve.

Value

A line with the true variogram model is added to the current plot on the graphics device. No values are returned.

Author(s)

Paulo J. Ribeiro Jr. paulojus@leg.ufpr.br,
Peter J. Diggle p.diggle@lancaster.ac.uk.

References

Further information on the package geoR can be found at:
http://www.leg.ufpr.br/geoR.

See Also

lines.variomodel, grf, plot.grf, curve.

Examples

sim <- grf(100, cov.pars=c(1, .25)) # simulates data
plot(variog(sim, max.dist=1))       # plot empirical variogram

[Package geoR version 1.7-5.2.1 Index]