lines.variomodel.likGRF {geoR}	R Documentation

Adds a Variogram Line to a Variogram Plot

Description

This function adds a fitted variogram based on the estimates of the model parameters returned by the function likfit to a current variogram plot.

Usage

## S3 method for class 'likGRF'
lines.variomodel(x, max.dist, scaled = FALSE, ...)

Arguments

x
an object of the class likGRF which is a list containing information about the fitted model parameters, typically an output of the function likfit.

max.dist
maximum distance (x-axis) to compute and draw the line representing the variogram model. The default is the distance given by obj$max.dist.

scaled
logical. If TRUE the total sill in the plot is equals to 1.

...
arguments to be passed to the function curve.

Details

Adds variogram model(s) to a plot. In conjuction with plot.variogram can be used to compare sample variograms against fitted models returned by likfit.

Value

A line with a variogram model is added to a plot on the current graphics device. No values are returned.

Author(s)

Paulo Justiniano Ribeiro Jr. paulojus@leg.ufpr.br,
Peter J. Diggle p.diggle@lancaster.ac.uk.

References

Further information on the package geoR can be found at:
http://www.leg.ufpr.br/geoR.

See Also

lines.variomodel, lines.variomodel.variofit, plot.variogram, lines.variogram, variofit, likfit, curve.

Examples

# compute and plot empirical variogram
vario <- variog(s100, max.dist = 1)
plot(vario)
# estimate parameters
vario.ml <- likfit(s100, ini = c(1, .3), fix.nugget = TRUE)
# adds fitted model to the plot  
lines(vario.ml)

[Package geoR version 1.7-5.2.1 Index]