loglik.GRF {geoR}	R Documentation

Log-Likelihood for a Gaussian Random Field

Description

This function computes the value of the log-likelihood for a Gaussian random field.

Usage

loglik.GRF(geodata, coords = geodata$coords, data = geodata$data,
           obj.model = NULL, cov.model = "exp", cov.pars, nugget = 0,
           kappa = 0.5, lambda = 1, psiR = 1, psiA = 0,
           trend = "cte", method.lik = "ML", compute.dists = TRUE,
           realisations = NULL)

Arguments

geodata
a list containing elements coords and data as described next. Typically an object of the class "geodata" - a geoR data-set. If not provided the arguments coords and data must be provided instead.

coords
an n x 2 matrix, each row containing Euclidean coordinates of the n data locations. By default it takes the element coords of the argument geodata.

data
a vector with data values. By default it takes the element data of the argument geodata.

obj.model
a object of the class variomodel with a fitted model. Tipically an output of likfit or variofit.

cov.model
a string specifying the model for the correlation function. For further details see documentation for cov.spatial.

cov.pars
a vector with 2 elements with values of the covariance parameters sigma^2 (partial sill) and phi (range parameter).

nugget
value of the nugget parameter. Defaults to 0.

kappa
value of the smoothness parameter. Defaults to 0.5.

lambda
value of the Box-Cox tranformation parameter. Defaults to 1.

psiR
value of the anisotropy ratio parameter. Defaults to 1, corresponding to isotropy.

psiA
value (in radians) of the anisotropy rotation parameter. Defaults to zero.

trend
specifies the mean part of the model. The options are: "cte" (constant mean), "1st" (a first order polynomial on the coordinates), "2nd" (a second order polynomial on the coordinates), or a formula of the type ~X where X is a matrix with the covariates (external trend). Defaults to "cte".

method.lik
options are "ML" for likelihood and "REML" for restricted likelihood. Defaults to "ML".

compute.dists
for internal use with other function. Don't change the default unless you know what you are doing.

realisations
optional. A vector indicating replication number for each data. For more details see as.geodata.

Details

The expression log-likelihood is:

l(theta) = -(n/2) * log(2 * pi) - 0.5 * log|V| - 0.5 * (y - F b)' V^{-1} (y - F b),

where n is the size of the data vector y, b is the mean (vector) parameter with dimention p, V is the covariance matrix and F is the matrix with the values of the covariates (a vector of 1's if the mean is constant.

The expression restricted log-likelihood is:

rl(θ) = -((n-p)/2) * log (2 * pi) + 0.5 * log |F'F| - 0.5 * log |V| - 0.5 * log |F'VF| - 0.5 * (y - Fb)' V^(-1) (y - Fb).

Value

The numerical value of the log-likelihood.

Author(s)

Paulo Justiniano Ribeiro Jr. paulojus@leg.ufpr.br,
Peter J. Diggle p.diggle@lancaster.ac.uk.

References

Further information on the package geoR can be found at:
http://www.leg.ufpr.br/geoR.

See Also

likfit for likelihood-based parameter estimation.

Examples

loglik.GRF(s100, cov.pars=c(0.8, .25), nugget=0.2)
loglik.GRF(s100, cov.pars=c(0.8, .25), nugget=0.2, met="RML")

## Computing the likelihood of a model fitted by ML
s100.ml <- likfit(s100, ini=c(1, .5))
s100.ml
s100.ml$loglik
loglik.GRF(s100, obj=s100.ml)

## Computing the likelihood of a variogram fitted model
s100.v <- variog(s100, max.dist=1)
s100.vf <- variofit(s100.v, ini=c(1, .5))
s100.vf
loglik.GRF(s100, obj=s100.vf)

[Package geoR version 1.7-5.2.1 Index]