names.geodata {geoR}	R Documentation

Lists names of the key elements of a geodata object

Description

Produces a list with the names of the main elements of geodata object: coords, data, units.m, covariate and realisation. Can be useful to list names before using {subset.geodata}.

Usage

## S3 method for class 'geodata'
names(x)

Arguments

x
an object of the class geodata.

Value

A list with

coords
names of the coordinates in the geodata object.

data
name(s) of the data elements in the geodata object.

units.m
returns the string units.m.

covariates
return the covariate(s) name(s) if present in the geodata object

realisations
returns the string units.m if present in the geodata object.

other
other elements in the geodata object.

See Also

names, subset.geodata, as.geodata.

Examples

names(ca20)

[Package geoR version 1.7-5.2.1 Index]