nearloc {geoR}	R Documentation

Near location to a point

Description

For a given set of points and locations identified by 2D coordinates this function finds the nearest location of each point

Usage

nearloc(points, locations, positions = FALSE)

Arguments

points
a matrix, data-frame or list with the 2D coordinates of a set of points for which you want to find the nearest location.

locations
a matrix, data-frame or list with the 2D coordinates of a set of locations.

positions
logical defining what to be returned. If TRUE the function returns the positions of the locations, otherwise the coordinates of the locations. Defaults to FALSE.

Value

If positions = FALSE the function returns a matrix, data-frame or list of the same type and size as the object provided in the argument points with the coordinates of the nearest locations.

If positions = FALSE the function returns a vector with the position of the nearest points in the locations object.

Author(s)

Paulo J. Ribeiro Jr. paulojus@leg.ufpr.br,
Peter J. Diggle p.diggle@lancaster.ac.uk.

See Also

loccoords

Examples

set.seed(276)
gr <- expand.grid(seq(0,1, l=11), seq(0,1, l=11))
plot(gr, asp=1)
pts <- matrix(runif(10), nc=2)
points(pts, pch=19)
near <- nearloc(points=pts, locations=gr)
points(near, pch=19, col=2)
rownames(near)
nearloc(points=pts, locations=gr, pos=TRUE)

[Package geoR version 1.7-5.2.1 Index]