.nlmP {geoR}	R Documentation

Adapts nlm for Constraints in the Parameter Values

Description

This function adapts the R function nlm to allow for constraints (upper and/or lower bounds) in the values of the parameters.

Usage

.nlmP(objfunc, params, lower=rep(-Inf, length(params)),
      upper=rep(+Inf, length(params)), ...)

Arguments

objfunc
the function to be minimized.

params
starting values for the parameters.

lower
lower bounds for the variables. Defaults to -Inf.

upper
upper bounds for the variables. Defaults to -Inf.

...
further arguments to be passed to the function nlm.

Details

Constraints on the parameter values are internally imposed by using exponential, logarithmic, and logit transformation of the parameter values.

Value

The output is the same as for the function nlm.

Author(s)

Patrick E. Brown p.brown@lancaster.ac.uk.
Adapted and included in geoR by
Paulo Justiniano Ribeiro Jr. paulojus@leg.ufpr.br

References

Further information on the package geoR can be found at:
http://www.leg.ufpr.br/geoR.

See Also

nlm, optim.

[Package geoR version 1.7-5.2.1 Index]