parana {geoR}	R Documentation

Rainfall Data from Parana State, Brasil

Description

This data-set was used by Diggle and Ribeiro (2001) to illustrate the methods discussed in the paper. The data reported analysis was carried out using the package geoR.

The data refers to average rainfall over different years for the period May-June (dry-season). It was collected at 143 recording stations throughout Parana State, Brasil.

Usage

data(parana)

Format

The object parana of the class geodata, which is a list containing the following components:

coords

a matrix with the coordinates of the recording stations.

data

a vector with the average recorded rainfall for the May-June period.

borders

a matrix with the coordinates defining the borders of Parana state.

loci.paper

a matrix with the coordinates of the four prediction locations discussed in the paper.

Source

The data were collected at several recording stations at Parana State, Brasil, belonging to the following companies: COPEL, IAPAR, DNAEE, SUREHMA and INEMET.

The data base was organized by Laura Regina Bernardes Kiihl (IAPAR, Instituto Agronomico do Parana, Londrina, Brasil) and the fraction of the data included in this data-set was provided by Jacinta Loudovico Zamboti (Universidade Estadual de Londrina, Brasil). The coordinates of the borders of Parana State were provided by Joao Henrique Caviglione (IAPAR).

References

Diggle, P.J. \& Ribeiro Jr, P.J. (2002) Bayesian inference in Gaussian model-based geostatistics. Geographical and Environmental Modelling, Vol. 6, No. 2, 129-146.

Examples

summary(parana)
plot(parana, bor=borders)

[Package geoR version 1.7-5.2.1 Index]