pars.limits {geoR}	R Documentation

Set limits for the parameter values

Description

The functions likfit and variofit in the package geoR

Usage

pars.limits(phi = c(lower = 0, upper = +Inf),
            sigmasq = c(lower = 0, upper = +Inf),
            nugget.rel = c(lower = 0, upper = +Inf),
            kappa = c(lower = 0, upper = +Inf),
            kappa2 = c(lower = 0, upper = +Inf), 
            lambda = c(lower = -3, upper = 3),
            psiR = c(lower = 1, upper = +Inf),
            psiA = c(lower = 0, upper = 2 * pi),
            tausq.rel = nugget.rel)

Arguments

phi
a two elements vector with limits for the parameter phi. Defaults to [0, +Inf]

sigmasq
idem for sigmasq. Defaults to [0, +Inf]

nugget.rel
idem for nugget.rel. Defaults to [0, +Inf]

kappa,kappa2
idem. Defaults to [0, +Inf]

lambda
idem for lambda. Defaults to [-3, +3]. Only used in likfit.

psiR
idem for psiR. Defaults to [1, +Inf]. Only used in likfit.

psiA
idem for psiA. Defaults to [0, 2 pi]. Only used in likfit.

tausq.rel
idem for tausq.rel. Defaults to [0, +Inf]

Details

Lower and upper limits for parameter values can be individually specified. For example, including the following in the function call in likfit or variofit:
limits = pars.limits(phi=c(0, 10), lambda=c(-2.5, 2.5)),
will change the limits for the parameters phi and lambda. Default values are used if the argument limits is not provided.

Value

A list of a 2 elements vector with limits for each parameters

See Also

likfit, variofit

Examples

pars.limits(phi=c(0,10))
pars.limits(phi=c(0,10), sigmasq=c(0, 100))

[Package geoR version 1.7-5.2.1 Index]