plot.grf {geoR}	R Documentation

Plots Variograms for Simulated Data

Description

This function plots variograms for simulated geostatistical data generated by the function grf.

Usage

## S3 method for class 'grf'
plot(x, model.line = TRUE, plot.locations = FALSE, ...)

Arguments

x
an object of the class grf, typically an output of the function grf.

model.line
logical. If TRUE the true variogram model is added to the plot with the sample variogram(s).

plot.locations
logical. If TRUE a plot with data locations is also shown.

...
further arguments to be passed to the functions variog and plot.

Value

A plot with the empirical variogram(s) is produced on the output device. No values are returned.

Author(s)

Paulo Justiniano Ribeiro Jr. paulojus@leg.ufpr.br,
Peter J. Diggle p.diggle@lancaster.ac.uk.

References

Further information on the package geoR can be found at:
http://www.leg.ufpr.br/geoR.

See Also

grf for simulation of Gaussian random fields, plot.variogram for plotting empirical variogram, variog for computation of empirical variograms and plot for the generic plotting function.

Examples

op <- par(no.readonly = TRUE)
par(mfrow=c(2,1))
sim1 <- grf(100, cov.pars=c(10, .25))
# generates simulated data
plot(sim1, plot.locations = TRUE)
#
# plots the locations and the sample true variogram model
#
par(mfrow=c(1,1))
sim2 <- grf(100, cov.pars=c(10, .25), nsim=10)
# generates 10 simulated data
plot(sim1)
# plots sample variograms for all simulations with the true model
par(op)

[Package geoR version 1.7-5.2.1 Index]