plot.krige.bayes {geoR}	R Documentation

Plots Prior and/or Posterior Distributions

Description

Produces plots the priors and posteriors distribuitions for the paramters phi and tausq.rel based on results returned by krige.bayes.

Usage

## S3 method for class 'krige.bayes'
plot(x, phi.dist = TRUE, tausq.rel.dist = TRUE, add = FALSE,
                 type=c("bars", "h", "l", "b", "o", "p"), thin, ...)

Arguments

x
an object of the class krige.bayes, with an output of the funtions krige.bayes.

phi.dist
logical indicating whether or not plot the distributions for this parameter.

tausq.rel.dist
logical indicating whether or not plot the distributions for this parameter.

add
logical. If TRUE plots is added to current one.

type
indicates the type of plot. Option "bars" uses the function barplot and the others uses matplot.

thin
a numerical vector defining the thining for values of the parameters phi and tausq.rel respectively. This improves visualisation when there are many values in the discrete distribution of the parameters.

...
further arguments for the plotting function.

Value

For plot.krige.bayes a plot is produced or added to the current graphics device. No values are returned.

Author(s)

Paulo J. Ribeiro Jr. paulojus@leg.ufpr.br,
Peter J. Diggle p.diggle@lancaster.ac.uk.

See Also

krige.bayes, barplot, matplot.

Examples

## See documentation for krige.bayes

[Package geoR version 1.7-5.2.1 Index]