plot.proflik {geoR}	R Documentation

Plots Profile Likelihoods

Description

This function produces plots of the profile likelihoods computed by the function proflik.

Usage

## S3 method for class 'proflik'
plot(x, pages = c("user", "one", "two"), uni.only, bi.only,
             type.bi = c("contour", "persp"), conf.int = c(0.90, 0.95),
             yaxis.lims = c("conf.int", "as.computed"),
             by.col = TRUE, log.scale = FALSE, use.splines = TRUE,
             par.mar.persp = c(0, 0, 0, 0), ask = FALSE, ...)

Arguments

x
an object of the class proflik, typically an output of the function proflik.

pages
specify how the plots will be arranged in the graphics device. The default option, "user", uses the current graphical parameters. The option "one" places all the profiles in the same page and the option "two" places the univariate profiles in one page and the bivariate profiles in a second page.

uni.only
only 1-D profiles are plotted even if the object contains results about the 2-D profiles.

bi.only
only 2-D profile are plotted even if the object contains results about the 1-D profiles.

type.bi
Type of plot for the 2-D profiles. Options are "contour" for contour plot (the default) and "persp" for perspective plot.

conf.int
a vector with numbers in the interval [0,1] specifying levels of the (approximated) confidence intervals. Defaults corresponds to the levels 90% and 95%.

yaxis.lims
defines the lower limits for the y-axis in the 1-D plots. If "conf.int" the limit is determined by the level of the confidence interval (the default) otherwise will be determined by the smallest computed value.

by.col
logical, If TRUE the plots are arranged by columns in a multiple graphics device.

log.scale
plots the x-axis in the logarithmic scale. Defaults to FALSE.

use.splines
logical. If TRUE (the default) the function spline is used to interpolate between the points computed by proflik.

par.mar.persp
graphical parameters to be used with persp plots. For more details see par.

ask
logical. Defines whether or not the user is prompted before each plot is produced.

...
additional arguments to be passed to the functions plot, contour and/or persp.

Value

Produces plots with the profile likelihoods on the current graphics device. No values are returned.

Author(s)

Paulo Justiniano Ribeiro Jr. paulojus@leg.ufpr.br,
Peter J. Diggle p.diggle@lancaster.ac.uk.

References

Further information on the package geoR can be found at:
http://www.leg.ufpr.br/geoR.

See Also

proflik for computation of the profile likelihoods. For the generic plotting functions see plot, contour, persp. See spline for interpolation.

Examples

# see examples in the documentation for the function proflik()

[Package geoR version 1.7-5.2.1 Index]