plot.variog4 {geoR}	R Documentation

Plot Directional Variograms

Description

This function plot directional variograms computed by the function variog4. The omnidirectional variogram can be also included in the plot.

Usage

## S3 method for class 'variog4'
plot(x, omnidirectional=FALSE, same.plot=TRUE, legend = TRUE, ...)

Arguments

x
an object of the class variog4, typically an output of the function variog4.

omnidirectional
logical. Indicates whether the omnidirectional variogram is included in the plot.

same.plot
logical. Indicates whether the directional variograms are plotted in the same or separated plots.

legend
logical indicating whether legends are automatically included in the plots.

...
further arguments to be passed to the function plot. Typical arguments are col, lty, lwd. For same.plot = TRUE the arguments are passed to the function matplot which is used to produce the plot.

Value

A plot is produced on the output device. No values returned.

Author(s)

Paulo J. Ribeiro Jr. paulojus@leg.ufpr.br,
Peter J. Diggle p.diggle@lancaster.ac.uk.

References

Further information about the geoR package can be found at:
http://www.leg.ufpr.br/geoR.

See Also

variog4 for variogram calculations and matplot for multiple lines plotting.

Examples

s100.v4 <- variog4(s100, max.dist=1)
# Plotting variograms for the four directions
plot(s100.v4)
# changing plot options
plot(s100.v4, lwd=2)
plot(s100.v4, lty=1, col=c("darkorange", "darkblue", "darkgreen","darkviolet"))
plot(s100.v4, lty=1, lwd=2)
# including the omnidirectional variogram
plot(s100.v4, omni=TRUE)
# variograms on different plots
plot(s100.v4, omni=TRUE, same=FALSE)

[Package geoR version 1.7-5.2.1 Index]