plot.variogram {geoR}	R Documentation

Plot Empirical Variogram

Description

Plots sample (empirical) variogram computed using the function variog.

Usage

## S3 method for class 'variogram'
plot(x, max.dist, vario.col = "all", scaled = FALSE,
               var.lines = FALSE, envelope.obj = NULL,
               pts.range.cex, bin.cloud = FALSE,  ...)

Arguments

x
an object of the class "variogram", typically an output of the function variog.

max.dist
maximum distance for the x-axis. The default is the maximum distance for which the sample variogram was computed.

vario.col
only used if obj has information on more than one empirical variogram. The default "all" indicates that variograms of all variables should be plotted. Alternativelly a numerical vector can be used to select variables.

scaled
If TRUE the variogram values are divided by the sample variance. This allows comparison of variograms of variables measured in different scales.

var.lines
If TRUE a horizontal line is drawn at the value of the variance of the data (if scaled = F) or at 1 (if scaled = T).

envelope.obj
adds a variogram envelope computed by the function variog.model.env or variog.mc.env.

pts.range.cex
optional. A two elements vector with maximum and minimum values for the caracter expansion factor cex. If provided the point sizes in binned variogram are proportional to the number of pairs of points used to compute each bin.

bin.cloud
logical. If TRUE and the sample variogram was computed with the option bin.cloud = TRUE, box-plots of values at each bin are plotted instead of the empirical variograms.

...
other arguments to be passed to the function plot or matplot

Details

This function plots empirical variograms. Toghether with lines.variogram can be used to compare sample variograms of different variables and to compare variogram models against the empirical variogram.

It uses the function matplot when plotting variograms for more them one variable.

Value

Produces a plot with the sample variogram on the current graphics device. No values are returned.

Author(s)

Paulo Justiniano Ribeiro Jr. paulojus@leg.ufpr.br,
Peter J. Diggle p.diggle@lancaster.ac.uk

References

Further information on the package geoR can be found at:
http://www.leg.ufpr.br/geoR.

See Also

variog for variogram calculations, lines.variogram and lines.variomodel for adding lines to the current plot, variog.model.env and variog.mc.env for variogram envelops computation, matplot for multiple lines plot and plot for generic plot function.

Examples

op <- par(no.readonly = TRUE)
sim <- grf(100, cov.pars=c(1, .2)) # simulates data
vario <- variog(sim, max.dist=1)   # computes sample variogram
par(mfrow=c(2,2))
plot(vario)                     # the sample variogram
plot(vario, scaled = TRUE)      # the scaled sample variogram
plot(vario, max.dist = 1)       # limiting the maximum distance
plot(vario, pts.range = c(1,3)) # points sizes proportional to number of pairs
par(op)

[Package geoR version 1.7-5.2.1 Index]