practicalRange {geoR}	R Documentation

Pratical range for correlation functions

Description

Computes practical ranges for the correlation functions implemented in the geoR package

Usage

practicalRange(cov.model, phi, kappa = 0.5, correlation = 0.05, ...)

Arguments

cov.model
correlation model as documented in cov.spatial.

phi
correlation parameter as documented in cov.spatial.

kappa
additional correlation parameter as documented in cov.spatial.

correlation
correlation threshold for asymptotic models. Defaults to 0.05.

...
arguments to be passed to optimise.

Value

A scalar with the value of the practical range.

See Also

cov.spatial

Examples

practicalRange("exp", phi=10)
practicalRange("sph", phi=10)
practicalRange("gaus", phi=10)
practicalRange("matern", phi=10, kappa=0.5)
practicalRange("matern", phi=10, kappa=1.5)
practicalRange("matern", phi=10, kappa=2.5)

[Package geoR version 1.7-5.2.1 Index]