pred_grid {geoR}	R Documentation

Generates a 2D Prediction Grid

Description

This function facilitates the generation of a 2D prediction grid for geostatistical kriging.

Usage

pred_grid(coords, y.coords = NULL, ..., y.by = NULL,
          y.length.out = NULL, y.along.with = NULL)

Arguments

coords
a list, matrix or data-frame with xy-coordinates of prediction points or a vector with x-coordinates.

y.coords
a vector with y-coordinates. Needed if argument coords provides only x-coordinates.

...
arguments by or length.out to be passed to the function rep. These arguments are used for the x-coordinates and are default optioons for y-coordinates.

y.by
Optional. by argument for rep to be used with the y-coordinates.

y.length.out
Optional. length.out argument for rep to be used with the y-coordinates.

y.along.with
Optional. along.with argument for rep to be used with the y-coordinates.

Value

An two column data-frame which is on output of expand.grid.

See Also

See seq and expand.grid which are used internally and locations.inside and polygrid to select points inside a border.

Examples

pred_grid(c(0,1), c(0,1), by=0.25) ## create a grid in a unit square
loc0 <- pred_grid(ca20$borders, by=20)
points(ca20, borders=borders)
points(loc0, pch="+")
points(locations.inside(loc0, ca20$border), pch="+", col=2)

[Package geoR version 1.7-5.2.1 Index]