predict.BGCCM {geoR}	R Documentation

Prediction for the bivariate Gaussian common component geostatistical model

Description

Performs prediction for the bivariate Gaussian common component geostatistical model

Usage

## S3 method for class 'BGCCM'
predict(object, locations, borders,
              variable.to.predict = 1, ...)

Arguments

object
on object of the class BGCCMfit, which is an output of likfitBGCCM.

locations
an N x 2 matrix or data-frame with the 2-D coordinates of the N prediction locations, or a list for which the first two components are used. Input is internally checked by the function check.locations.

borders
optional. If missing, by default reads the element borders of the geodata object of the variable to be predicted. Ignored if set to NULL. If a two column matrix defining a polygon is provided the prediction is performed only at locations inside this polygon.

variable.to.predict
scalar with options for values or 2 indicating which variable is to be predicted.

...
not yet used.

Value

A list of the class BGCCMpred with components:

predicted
predicted values.

krige.var
prediction variances.

Warning

This is a new function and still in draft format and pretty much untested.

Author(s)

Paulo Justiniano Ribeiro Jr. paulojus@leg.ufpr.br,
Peter J. Diggle p.diggle@lancaster.ac.uk.

See Also

likfitBGCCM

Examples

# see http://www.leg.ufpr.br/geoR/tutorials/CCM.R

[Package geoR version 1.7-5.2.1 Index]