print.BGCCM {geoR}	R Documentation

Prints an summary of of the output from likfitBGCCM.

Description

Prints a short version of an object of the class BGCCM.

Usage

## S3 method for class 'BGCCM'
print(x, ...)

Arguments

x
an object of the class BGCCM.

...
arguments to be passed to format.

See Also

format for options to format the output.

[Package geoR version 1.7-5.2.1 Index]