read.geodata {geoR}	R Documentation

Reads and Converts Data to geoR Format

Description

Reads data from a ASCII file and converts it to an object of the class geodata, the standard data format for the geoR package.

Usage

read.geodata(file, header = FALSE, coords.col = 1:2, data.col = 3,
             data.names = NULL, covar.col = NULL, covar.names = "header",
             units.m.col = NULL, realisations = NULL,
             na.action = c("ifany", "ifdata", "ifcovar", "none"),
             rep.data.action, rep.covar.action, rep.units.action, ...)

Arguments

file
a string with the name of the ASCII file.

header
logical. Indicates whether the variables names should be read from the first line of the input file.

coords.col
a vector with the numbers of the columns containing the coordinates.

data.col
a scalar or vector with the number of the column(s) containing the data.

data.names
a string or vector of strings with names for the data columns. Only valid if there is more than one column of data. By default the names in the original object are used.

covar.col
optional. A scalar or vector with the number of the column(s) with the values of the covariate(s).

covar.names
optional. A vector with the names of the the covariates. By default the names in the original object are used.

units.m.col
optional. A scalar with the column number corresponding to the offset variable. Alternativelly can be a character vector with the name of the offset. This option is particularly relevant when using the package geoRglm.

realisations
optional. A vector indicating the replication number. For more details see documentation for as.geodata.

na.action
a string. Defines action to be taken in the presence of NA's. For more details see documentation for as.geodata.

rep.data.action
a string or a function. Defines action to be taken when there is more than one data at the same location. For more details see documentation for as.geodata.

rep.covar.action
a string or a function. Defines action to be taken when there is more than one covariate at the same location. For more details see documentation for as.geodata.

rep.units.action
a string or a function. Defines action to be taken on the element units.m, if present when there is more than one data at the same location. The default option is the same value set for rep.data.action.

...
further arguments to be passed to the function read.table.

Details

The function read.table is used to read the data from the ASCII file and then as.geodata is used to convert to an object of the class geodata.

Value

An object of the class geodata. See documentation for the function as.geodata for further details.

Author(s)

Paulo Justiniano Ribeiro Jr. paulojus@leg.ufpr.br,
Peter J. Diggle p.diggle@lancaster.ac.uk.

References

Further information on the package geoR can be found at:
http://www.leg.ufpr.br/geoR.

See Also

as.geodata to convert existing R objects, read.table, the basic R function used to read ASCII files, and list for detailed information about lists.

[Package geoR version 1.7-5.2.1 Index]