s100 and s121 {geoR}	R Documentation

Simulated Data-Sets which Illustrate the Usage of the Package geoR

Description

These two simulated data sets are the ones used in the Technical Report which describes the package geoR (see reference below). These data-sets are used in several examples throughout the package documentation.

Usage

data(s100)

data(s121)

Format

Two objects of the class geodata. Both are lists with the following components:

coords

the coordinates of data locations.

data

the simulated data. Notice that for s121 this a 101x10 matrix with 10 simulations.

cov.model

the correlation model.

nugget

the values of the nugget parameter.

cov.pars

the covariance parameters.

kappa

the value of the parameter kappa.

lambda

the value of the parameter lambda.

References

Ribeiro Jr, P.J. and Diggle, P.J. (1999) geoS: A geostatistical library for S-PLUS. Technical report ST-99-09, Dept of Maths and Stats, Lancaster University.

Further information on the package geoR can be found at:
http://www.leg.ufpr.br/geoR.

Examples

plot(s100)
plot(s121, type="l")

[Package geoR version 1.7-5.2.1 Index]