sample.prior {geoR}	R Documentation

Sample the prior distribution

Description

Sample quadruples (beta, sigma^2, phi, tau^2.rel) from the prior distribution of parameters specifying a Gaussian random field. Typically the prior is specified in the same manner as when calling krige.bayes.

Usage

sample.prior(n, kb.obj=NULL, prior=prior.control())

Arguments

n
number of samples

kb.obj
on object with an output of krige.bayes.

prior
an call to prior.control. Unnecessary if kb.obj is provided.

Value

A p+3 x 4 data-frame with a sample of the prior distribution of model parameters, where p is the length of the mean parameter beta.

Author(s)

Paulo J. Ribeiro Jr. paulojus@leg.ufpr.br,
Peter J. Diggle p.diggle@lancaster.ac.uk.

References

Further information on the package geoR can be found at:
http://www.leg.ufpr.br/geoR.

See Also

krige.bayes and sample.posterior.

Examples

sample.prior(50, 
             prior=prior.control(beta.prior = "normal", beta = .5, beta.var.std=0.1, 
                                 sigmasq.prior="sc", sigmasq=1.2, df.sigmasq= 2, 
                                 phi.prior="rec", phi.discrete = seq(0,2, l=21)))

[Package geoR version 1.7-5.2.1 Index]