set.coords.lims {geoR}	R Documentation

Sets Limits to Scale Plots

Description

This is an function typically called by functions in the package geoR to set limits for the axis when plotting spatial data.

Usage

set.coords.lims(coords, borders = coords, xlim, ylim, ...)

Arguments

coords
an nx2 matrix with coordinates.

borders
an nx2 matrix with coordinates.

xlim, ylim
the ranges to be encompassed by the x and y axes.

...
not used, included just for internal handling.

Value

A 2 \times 2 matrix with limits for the axis.

Author(s)

Paulo J. Ribeiro Jr. paulojus@leg.ufpr.br,
Peter J. Diggle p.diggle@lancaster.ac.uk.

[Package geoR version 1.7-5.2.1 Index]