soil250 {geoR}	R Documentation

Soil chemistry properties data set

Description

Several soil chemistry properties measured on a regular grid with 10x25 points spaced by 5 meters.

Usage

data(soil250)

Format

A data frame with 250 observations on the following 22 variables.

Linha

x-coordinate

Coluna

y-coordinate

Cota

elevation

AGrossa

a numeric vecto, sand portion of the sample.

Silte

a numeric vector, silt portion of the sample.

Argila

a numeric vector, sand portion of the sample.

pHAgua

a numeric vector, soil pH at water

pHKCl

a numeric vector, soil pH by KCl

Ca

a numeric vector, calcium content

Mg

a numeric vector, magnesium content

K

a numeric vector, potassio content

Al

a numeric vector, aluminium content

H

a numeric vector, hidrogen content

C

a numeric vector, carbon content

N

a numeric vector, nitrogen content

CTC

a numeric vector, catium exchange capability

S

a numeric vector, enxofrar content

V

a numeric vector

M

a numeric vector

NC

a numeric vector

CEC

a numeric vector

CN

a numeric vector, carbon/nitrogen relation

Details

Uniformity trial with 250 undisturbed soil samples collected at 25cm soil depth of spacing of 5 meters, resulting on a regular grid of 25 x 10 points.

See also the data-set wrc with other variables colected at the same points.

Source

Bassoi thesis

References

Bassoi papers

Examples

data(soil250)
ctc <- as.geodata(soil250, data.col=16)
plot(ctc)

[Package geoR version 1.7-5.2.1 Index]