soja98 {geoR}	R Documentation

Soya bean production and other variables in a uniformity trial

Description

Data on soya bean production in a uniformity trial measured at plots of size 5x5 meters and other soil properties measured in points given by the data coordinates.

Usage

data(soja98)

Format

A data frame with 256 observations on the following 10 variables.

X

a numeric vector with X-coordinates of the plot centres.

Y

a numeric vector with X-coordinates of the plot centres.

P

a numeric vector, phosphorous content.

PH

a numeric vector, soil pH.

K

a numeric vector, potassium content. (Cmol/dm^3)

MO

a numeric vector, organic matter. (percentage)

SB

a numeric vector, basis saturation.

iCone

a numeric vector, cone index, measuring mechanic resistence of the soil. (kg/cm^2)

Rend

a numeric vector, total soya production within the plot (kg).

PROD

a numeric vector, production converted to productivity by a unit of area - hectare (ton/ha).

Source

Souza, E.G.; Jojann, J. A.; Rocha, J. V.; Ribeiro, S. R. A.; Silva, M. S., Upazo, M. A. U.; Molin, J. P.; Oliveira, E. F.; Nóbrega, L. H. P. (1999) Variabilidade espacial dos atributos químicos do solo em um latossolo roxo distrófico da região de Cascavel-PR. Engenharia Agrícola. Jaboticabal, volume 18, nr. 3, p.80-92.

Examples

data(soja98)
plot(soja98)
require(geoR)
prod98 <- as.geodata(soja98, coords.col=1:2, data.col=)
plot(prod98, low=TRUE)

[Package geoR version 1.7-5.2.1 Index]