statistics.predictive {geoR}	R Documentation

Summary statistics from predictive distributions

Description

Computes summaries based on simulations of predictive distribution returned by krige.bayes and krige.conv.

Usage

statistics.predictive(simuls, mean.var = TRUE, quantile, threshold,
                      sim.means, sim.vars)

Arguments

simuls
object with simulations from the predictive distribution

mean.var
Logical. Indicates whether or not to compute mean and variances of the simulations at each location.

quantile
defines quantile estimator. See documentation for output.control .

threshold
defines probability estimator. See documentation for output.control.

sim.means
Logical. Indicates whether or not to compute the mean of of the conditional simulations.

sim.vars
Logical. Indicates whether or not to compute the variances of the conditional simulations.

Value

A list with one ore more of the following components.

mean
mean at each prediction location.

variance
variance at each prediction location.

quantiles
quantiles, at each prediction location.

probabilities
probabilities, at each prediction location, of been below the provided threshold.

sim.means
vector with means of each conditional simulation.

sim.vars
vector with variances of each conditional simulation.

Author(s)

Paulo J. Ribeiro Jr. paulojus@leg.ufpr.br,
Peter J. Diggle p.diggle@lancaster.ac.uk.

References

Further information on the package geoR can be found at:
http://www.leg.ufpr.br/geoR.

[Package geoR version 1.7-5.2.1 Index]