subset.geodata {geoR}	R Documentation

Method for subsetting geodata objects

Description

Subsets a object of the class geodata by transforming it to a data-frame, using subset and back transforming to a geodata object.

Usage

## S3 method for class 'geodata'
subset(x, ..., other = TRUE)

Arguments

x
an object of the class geodata.

...
arguments to be passed to subset.data.frame.

other
logical. If TRUE non-standard geodata elements of the original geodata object are copied to the resulting object.

Value

A list which is an object of the class geodata.

See Also

subset for the generic function and methods and as.geodata for more information on geodata objects.

Examples

subset(ca20, data > 70)
subset(ca20, area == 1)

[Package geoR version 1.7-5.2.1 Index]