summary.geodata {geoR}	R Documentation

Summaries for geodata object

Description

Sumarises each of the main elements of an object of the class geodata.

Usage

## S3 method for class 'geodata'
summary(object, lambda =1, add.to.data = 0,
                by.realisations=TRUE, ...)

Arguments

object
an object of the class geodata.

lambda
value of the Box-Cox transformation parameter. Two particular cases are lambda = 1 which corresponds to no transformation and lambda = 0 corresponding to the log-transformation.

add.to.data
scalar, Constant value to be added to the data values. Only used if a value different from 1 is passed to the argument lambda.

by.realisations
logical. Indicates whether the summary must be performed separatly for each realisation, if the geodata object contains the element realisations. Defaults to TRUE.

...
further arguments to be passed to the function summary.default.

Value

A list with components

coords.summary
a matrix with minimum and maximum values for the coordinates.

distances.summary
minimum and maximum distances between pairs of points.

borders.summary
a matrix with minimum and maximum values for the coordinates. Only returned if there is an element borders in the geodata object.

data.summary
summary statistics (min, max, quartiles and mean) for the data.

units.m.summary
summary statistics (min, max, quartiles and mean) for the offset variable. Only returned if there is an element units.m in the geodata object.

covariate.summary
summary statistics (min, max, quartiles and mean) for the covariate(s). Only returned if there is an element covariate in the geodata object.

others
names of other elements if present in the geodata object.

Author(s)

Paulo Justiniano Ribeiro Jr. paulojus@leg.ufpr.br,
Peter J. Diggle p.diggle@lancaster.ac.uk.

References

Further information on the package geoR can be found at:
http://www.leg.ufpr.br/geoR.

See Also

summary, as.geodata.

Examples

summary(s100)

summary(ca20)

[Package geoR version 1.7-5.2.1 Index]