summary.variofit {geoR}	R Documentation

Summarize Results of Variogram Estimation

Description

This function prints a summary of the parameter estimation results given by variofit.

Usage

## S3 method for class 'variofit'
summary(object, ...)

Arguments

object
an object of the class "variomodel" typically an output of variofit.

...
other arguments to be passed to the function print or summary.

Value

Prints a summary of the estimation results on the screen or other output device.

Author(s)

Paulo J. Ribeiro Jr. paulojus@leg.ufpr.br,
Peter J. Diggle p.diggle@lancaster.ac.uk.

References

Further information on the package geoR can be found at:
http://www.leg.ufpr.br/geoR.

See Also

The functions variofit for variogram based estimation. For likelihood based parameter estimation see likfit.

Examples

s100.vario <- variog(s100, max.dist=1)
wls <- variofit(s100.vario, ini=c(.5, .5), fix.nugget = TRUE)
wls
summary(wls)

[Package geoR version 1.7-5.2.1 Index]