tce {geoR}	R Documentation

TCE concentrations in groundwater in a vertical cross section

Description

Measurements at 56 locations of concentration of trichloroethylene (TCE) in groundwater on a transect in a fine-sand superficial aquifer. Extract from Kitanidis' book.

Usage

data(tce)

Format

An object of the class geodata which is a list with the elements:

coords

coordinates of the data location (feet).

data

the data vector with measurements of the TCE concentration (ppb).

Source

Kitanidis, P.K. Introduction to geostatistics - applications in hidrology (1997). Cambridge University Press.

Examples

summary(tce)
summary(tce, lambda=0)
plot(tce)
points(tce)
points(tce, lambda=0)

[Package geoR version 1.7-4 Index]