trend.spatial {geoR}	R Documentation

Builds the Trend Matrix

Description

Builds the trend matrix in accordance to a specification of the mean provided by the user.

Usage

trend.spatial(trend, geodata, add.to.trend)

Arguments

trend
specifies the mean part of the model. See DETAILS below.

geodata
optional. An object of the class geodata as described in as.geodata.

add.to.trend
optional. Specifies aditional terms to the mean part of the model. See details below.

Details

The implicity model assumes that there is an underlying process with mean mu(x), where x = (x1, x2) denotes the coordinates of a spatial location. The argument trend defines the form of the mean and the following options are allowed:

"cte"the mean is assumed to be constant over the region, in which case mu(x) = mu. This is the default option.

"1st"the mean is assumed to be a first order polynomial on the coordinates:

mu(x) = beta0 + beta1*x1 + beta2*x2.

"2nd"the mean is assumed to be a second order polynomial on the coordinates:

μ(x)= beta0 + beta1*x1 + beta2*x2 + beta3*(x1)^2 + beta4*(x2)^2 + beta5*x1*x2.

~ modela model specification. See formula for further details on how to specify a model in R using formulas. Notice that the model term before the ~ is not necessary. Typically used to include covariates (external trend) in the model.

Denote by x_1 and x_2 the spatial coordinates. The following specifications are equivalent:

trend = "1st" and trend = ~ x1 + x2

trend = "2nd" and trend = ~ x1 + x2 + I(x1^2) + I(x2^2) + I(x1*x2)

Search path for covariates
Typically, functions in the package geoR which calls trend.spatial will have the arguments geodata, coords and data.

When the trend is specifed as trend = ~ model the terms included in the model will be searched for in the following path sequence (in this order):

in the users/session Global environment

in the session search path

as elements of the list geodata

as columns in a data-frame geodata$covariates

as columns in a data-frame geodata$data

in remainder of the session search path

The argument add.to.trend adds terms to what is specified in the argument trend. This seems redundant but allow specifications of the type: trend="2nd", add.trend=~other.covariates.

Value

An object of the class trend.spatial which is an n x p trend matrix, where n is the number of spatial locations and p is the number of mean parameters in the model.

Note

This is an auxiliary function typically called by other geoR functions.

Author(s)

Paulo J. Ribeiro Jr. paulojus@leg.ufpr.br,
Peter J. Diggle p.diggle@lancaster.ac.uk.

References

Further information on the package geoR can be found at:
http://www.leg.ufpr.br/geoR.

See Also

The section DETAILS in the documentation for likfit for more about the underlying model.

Examples

# a first order polynomial trend
trend.spatial("1st", sic.100)[1:5,]
# a second order polynomial trend
trend.spatial("2nd", sic.100)[1:5,]
# a trend with a covariate
trend.spatial(~altitude, sic.100)[1:5,]
# a first degree trend plus a covariate
trend.spatial(~coords+altitude, sic.100)[1:5,]
# with produces the same as
trend.spatial("1st", sic.100, add=~altitude)[1:5,]
# and yet another exemple
trend.spatial("2nd", sic.100, add=~altitude)[1:5,]

[Package geoR version 1.7-5.2.1 Index]