varcovBGCCM {geoR}	R Documentation

Covariance matrix for the bivariate Gaussian common component geostatistical model

Description

Covariance matrix for the bivariate Gaussian common component geostatistical model or its inverse, and optionally the determinant of the matrix.

Usage

varcovBGCCM(dists.obj, cov0.pars, cov1.pars, cov2.pars,
            cov0.model = "matern", cov1.model = "matern",
            cov2.model = "matern", kappa0 = 0.5, kappa1 = 0.5,
            kappa2 = 0.5, scaled = TRUE, inv = FALSE, det = FALSE)

Arguments

dists.obj
a vector with distance values

cov0.pars
covarianve paremeter values for the common component

cov1.pars
covariance parameter for the individual structure of the first variable

cov2.pars
covariance parameter for the individual structure of the second variable

cov0.model
character indicating a valid correlation model

cov1.model
character indicating a valid correlation model

cov2.model
character indicating a valid correlation model

kappa0
scalar

kappa1
scalar

kappa2
scalar

scaled
logical

inv
logical. If TRUE the inverse of the covariance matrix is returned instead.

det
logical. Optional, if TRUE the logarithm of the detarminant of the covariance matrix is returned as an attribute.

Value

A matrix which is the covariance matrix for the bivariate Gaussian common component geostatistical model or its inverse if inv=TRUE. If det=T the logarithm of the determinant of the matrix is also returned as an attribute named logdetS.

Warning

This is a new function and still in draft format and pretty much untested.

Author(s)

Paulo J. Ribeiro Jr. paulojus@leg.ufpr.br,
Peter J. Diggle p.diggle@lancaster.ac.uk.

See Also

cov.spatial, varcov.spatial

Examples

# see http://www.leg.ufpr.br/geoR/tutorials/CCM.R

[Package geoR version 1.7-5.2.1 Index]