variog.mc.env {geoR}	R Documentation

Envelops for Empirical Variograms Based on Permutation

Description

Computes envelops for empirical variograms by permutation of the data values on the spatial locations.

Usage

variog.mc.env(geodata, coords = geodata$coords, data = geodata$data,
              obj.variog, nsim = 99, save.sim = FALSE, messages) 

Arguments

geodata
a list containing elements coords and data as described next. Typically an object of the class "geodata" - a geoR data-set. If not provided the arguments coords and data must be provided instead.

coords
an n x 2 matrix, each row containing Euclidean coordinates of the n data locations. By default it takes the element coords of the argument geodata.

data
a vector with the data values. By default it takes the element data of the argument geodata.

obj.variog
an object of the class "variogram", typically an output of the function variog.

nsim
number of simulations used to compute the envelope. Defaults to 99.

save.sim
logical. Indicates whether or not the simulated data are included in the output. Defaults to FALSE.

messages
logical. If TRUE, the default, status messages are printed while the function is running.

Details

The envelops are obtained by permutation. For each simulations data values are randomly allocated to the spatial locations. The empirical variogram is computed for each simulation using the same lags as for the variogram originally computed for the data. The envelops are computed by taking, at each lag, the maximum and minimum values of the variograms for the simulated data.

Value

An object of the class "variogram.envelope" which is a list with the following components:

u
a vector with distances.

v.lower
a vector with the minimum variogram values at each distance in u.

v.upper
a vector with the maximum variogram values at each distance in u.

simulations
a matrix with simulated data. Only returned if save.sim = TRUE.

Author(s)

Paulo Justiniano Ribeiro Jr. paulojus@leg.ufpr.br,
Peter J. Diggle p.diggle@lancaster.ac.uk.

References

Further information on the package geoR can be found at:
http://www.leg.ufpr.br/geoR.

See Also

variog.model.env for envelops computed by from a model specification, variog for variogram calculations, plot.variogram and variog.mc.env for graphical output.

Examples

s100.vario <- variog(s100, max.dist=1)
s100.env <- variog.mc.env(s100, obj.var = s100.vario)
plot(s100.vario, envelope = s100.env)

[Package geoR version 1.7-5.2.1 Index]