variog.model.env {geoR}	R Documentation

Envelops for Empirical Variograms Based on Model Parameters

Description

Computes envelopes for a empirical variogram by simulating data for given model parameters.

Computes bootstrap paremeter estimates

Usage

variog.model.env(geodata, coords = geodata$coords, obj.variog,
                 model.pars, nsim = 99, save.sim = FALSE, messages)

boot.variofit(geodata, coords = geodata$coords, obj.variog,
              model.pars, nsim = 99, trace = FALSE, messages)

Arguments

geodata
a list containing element coords as described next. Typically an object of the class "geodata" - a geoR data-set. If not provided the argument coords must be provided instead.

coords
an n x 2 matrix, each row containing Euclidean coordinates of the n data locations. By default it takes the element coords of the argument geodata.

obj.variog
an object of the class "variogram", typically an output of the function variog.

model.pars
a list with model specification and parameter values. The input is typically an object of the class variomodel which is an output of likfit, variofit. The required components of the list are:

beta, the mean parameter. Defaults to zero.

cov.model, the covariance model. Defaults to "exponential".

cov.pars, the covariance parameters sigma^2 and phi.

kappa, the extra covariance parameters for some of the covariance models. Defaults to 0.5.

nugget, the error component variance. Defaults to zero.

estimator.type, the type of variogram estimator. Options for "classical" and "robust". Defaults to obj.variog$estimator.

nsim
number of simulations used to compute the envelopes. Defaults to 99.

save.sim
logical. Indicates whether or not the simulated data are included in the output. Defaults to FALSE.

trace
logical. If TRUE the fitted values for the bootstrap parameter estimation are printend while the function is running.

messages
logical. If TRUE, the default, status messages are printed while the function is running.

Details

The envelopes are computed assuming a (transformed) Gaussian random field model. Simulated values are generated at the data locations, given the model parameters. The empirical variogram is computed for each simulation using the same lags as for the original variogram of the data. The envelopes are computed by taking, at each lag, the maximum and minimum values of the variograms for the simulated data.

Value

An object of the class "variogram.envelope" which is a list with the components:

u
a vector with distances.

v.lower
a vector with the minimum variogram values at each distance in u.

v.upper
a vector with the maximum variogram values at each distance in u.

simulations
a matrix with the simulated data. Only returned if save.sim = TRUE.

Author(s)

Paulo Justiniano Ribeiro Jr. paulojus@leg.ufpr.br,
Peter J. Diggle p.diggle@lancaster.ac.uk.

References

Further information on the package geoR can be found at:
http://www.leg.ufpr.br/geoR.

See Also

variog.mc.env for envelops computed by using data permutation, variog for variogram calculations, plot.variogram and variog.mc.env for graphical output. The functions likfit, variofit are used to estimate the model parameters.

Examples

s100.ml <- likfit(s100, ini = c(0.5, 0.5), fix.nugget = TRUE)
s100.vario <- variog(s100, max.dist = 1)
s100.env <- variog.model.env(s100, obj.v = s100.vario,
                             model.pars = s100.ml)
plot(s100.vario, env = s100.env)

[Package geoR version 1.7-5.2.1 Index]