wolfcamp {geoR}	R Documentation

Wolfcamp Aquifer Data

Description

Piezometric head measurements taken at the Wolfcamp Aquifer, Texas, USA. See Cressie (1993, p.212–214) for description of the scientific problem and the data. Original data were converted to SI units: coordinates are given in kilometers and pressure heads to meters.

Usage

data(wolfcamp)

Format

An object of the class "geodata", which is list with two components:

coords

the coordinates of the data locations. The distance are given in kilometers.

data

values of the piezometric head. The unit is heads to meters.

Source

Harper, W.V and Furr, J.M. (1986) Geostatistical analysis of potentiometric data in the Wolfcamp Aquifer of the Palo Duro Basin, Texas. Technical Report BMI/ONWI-587, Bettelle Memorial Institute, Columbus, OH.

References

Cressie, N.A.C (1993) Statistics for Spatial Data. New York: Wiley.

Papritz, A. and Moyeed, R. (2001) Parameter uncertainty in spatial prediction: checking its importance by cross-validating the Wolfcamp and Rongelap data sets. GeoENV 2000: Geostatistical for Environmental Applications. Ed. P. Monestiez, D. Allard, R. Froidevaux. Kluwer.

Examples

summary(wolfcamp)
plot(wolfcamp)

[Package geoR version 1.7-4 Index]