geoR-internal {geoR}	R Documentation

geoR internal functions

Description

These are functions internally called by other functions in the package geoR and not meant to be called by the user.

Author(s)

Paulo J. Ribeiro Jr. paulojus@leg.ufpr.br,
Peter J. Diggle p.diggle@lancaster.ac.uk.

[Package geoR version 1.7-5.2.1 Index]