globalvar {geoR}	R Documentation

Computes global variance

Description

Global variance computation for a set of locations using the covarianve model

Usage

globalvar(geodata, locations, coords = geodata$coords, krige)

Arguments

geodata
an object of the class geodata

locations
n by 2 matrix with a set of locations, typically a prediction grid

coords
data coordinates

krige
a list defining the model components and the type of kriging. It can take an output to a call to krige.control or a list with elements as for the arguments in krige.control.

Value

An scalar with the value of the global variance

Author(s)

Paulo Justiniano Ribeiro Jr. paulojus@leg.ufpr.br,
Peter J. Diggle p.diggle@lancaster.ac.uk.

References

Isaaks, E.S and Srivastava, R.M. (1989) An Introduction to Applied Geostatistics, pag. 508, eq. 20.7. Oxford University Press.

See Also

krige.conv for the kriging algorithm.

[Package geoR version 1.7-5.2.1 Index]