isaaks {geoR}	R Documentation

Data from Isaaks and Srisvastava's book

Description

Toy example used in the book An Introduction to Geostatistics to illustrate the effects of different models and parameters in the kriging results when predicting at a given point.

Usage

data(isaaks)

Format

An object of the class geodata which is a list with the elements:

coords

coordinates of the data location.

data

the data vector.

x0

coordinate of the prediction point.

Source

Isaaks, E.H. & Srisvastava, R.M. (1989) An Introduction to Applied Geostatistics. Oxford University Press.

Examples

isaaks
summary(isaaks)
plot(isaaks$coords, asp=1, type="n")
text(isaaks$coords, as.character(isaaks$data))
points(isaaks$x0, pch="?", cex=2, col=2)

[Package geoR version 1.7-5.2.1 Index]