lines.variogram.envelope {geoR}	R Documentation

Adds Envelopes Lines to a Variogram Plot

Description

Variogram envelopes computed by variog.model.env or variog.mc.env are added to the current variogram plot.

Usage

## S3 method for class 'variogram.envelope'
lines(x, lty = 3, ...)

Arguments

x
an object of the class "variogram.envelope", typically an output of the functions variog.model.env or variog.mc.env.

lty
line type. Defaults to 3.

...
arguments to be passed to the function lines.

Value

Lines defining the variogram envelope are added to the plotin the current graphics device.

Author(s)

Paulo Justiniano Ribeiro Jr. paulojus@leg.ufpr.br,
Peter J. Diggle p.diggle@lancaster.ac.uk.

References

Further information on the package geoR can be found at:
http://www.leg.ufpr.br/geoR.

See Also

variog for variogram computation, variog.model.env and variog.mc.env for computation of variogram envelopes, and lines for the generic function.

Examples

s100.vario <- variog(s100, max.dist = 1)
s100.ml <- likfit(s100, ini=c(.5, .5))
s100.mod.env <- variog.model.env(s100, obj.variog = s100.vario,
   model = s100.ml) 
s100.mc.env <- variog.mc.env(s100, obj.variog = s100.vario)
plot(s100.vario)
lines(s100.mod.env)
lines(s100.mc.env, lwd=2)

[Package geoR version 1.7-5.2.1 Index]