summary.likGRF {geoR}	R Documentation

Summarizes Parameter Estimation Results for Gaussian Random Fields

Description

Summarizes results returned by the function likfit.
Functions are methods for summary and print for the classes likGRF and summary.likGRF.

Usage

## S3 method for class 'likGRF'
summary(object, ...)
## S3 method for class 'likGRF'
print(x, digits = max(3, getOption("digits") - 3), ...)
## S3 method for class 'summary.likGRF'
print(x, digits = max(3, getOption("digits") - 3), ...)

Arguments

object
an object of class likGRF, typically a result of a call to likfit.

x
an object of class likGRF or class summary.likGRF, typically resulting from a call to likfit.

digits
the number of significant digits to use when printing.

...
extra arguments for print.

Details

A detailed summary of a object of the class likGRF is produced by by summary.likGRF and printed by print.summary.likGRF. This includes model specification with values of fixed and estimated parameters. A simplified summary of the parameter estimation is printed by print.likGRF.

Value

print.likGRF prints the parameter estimates and the value of the maximized likelihood.
summary.likGRF returns a list with main results of a call to likfit.
print.summary.likGRF prints these results on the screen (or other output device) in a "nice" format.

Author(s)

Paulo Justiniano Ribeiro Jr. paulojus@leg.ufpr.br,
Peter J. Diggle p.diggle@lancaster.ac.uk.

References

Further information on the package geoR can be found at:
http://www.leg.ufpr.br/geoR.

See Also

likfit, print, summary.

Examples

# See examples for the function likfit()

[Package geoR version 1.7-5.2.1 Index]