variog4 {geoR}	R Documentation

Computes Directional Variograms

Description

Computes directional variograms for 4 directions provided by the user.

Usage

variog4(geodata, coords = geodata$coords, data = geodata$data, 
        uvec = "default", breaks = "default", trend = "cte", lambda = 1,
        option = c("bin", "cloud", "smooth"),
        estimator.type = c("classical", "modulus"), 
        nugget.tolerance, max.dist, pairs.min = 2,
        bin.cloud = FALSE, direction = c(0, pi/4, pi/2, 3*pi/4), tolerance = pi/8,
        unit.angle = c("radians", "degrees"), messages, ...)

Arguments

geodata
a list containing element coords as described next. Typically an object of the class "geodata" - a geoR data-set. If not provided the arguments coords must be provided instead.

coords
an n x 2 matrix containing coordinates of the n data locations in each row. Defaults to geodata$coords, if provided.

data
a vector or matrix with data values. If a matrix is provided, each column is regarded as one variable or realization. Defaults to geodata$data, if provided.

uvec
a vector with values to define the variogram binning. For further details see documentation for variog.

breaks
a vector with values to define the variogram binning. For further details see documentation for variog.

trend
specifies the mean part of the model. The options are: "cte" (constant mean), "1st" (a first order polynomial on the coordinates), "2nd" (a second order polynomial on the coordinates), or a formula of the type ~X where X is a matrix with the covariates (external trend). Defaults to "cte".

lambda
values of the Box-Cox transformation parameter. Defaults to 1 (no transformation). If another value is provided the variogram is computed after transforming the data. A case of particular interest is lambda = 0 which corresponds to log-transformation.

option
defines the output type: the options "bin" returns values of binned variogram, "cloud" returns the variogram cloud and "smooth" returns the kernel smoothed variogram. Defaults to "bin".

estimator.type
"classical" computes the classical method of moments estimator. "modulus" returns the variogram estimator suggested by Hawkins and Cressie (see Cressie, 1993, pg 75). Defaults to "classical".

nugget.tolerance
a numeric value. Points which are separated by a distance less than this value are considered co-located. Defaults to zero.

max.dist
a numerical value defining the maximum distance for the variogram. Pairs of locations separated for distance larger than this value are ignored for the variogram calculation. Defaults to the maximum distance among the pairs of data locations.

pairs.min
a integer number defining the minimum numbers of pairs for the bins. For option = "bin", bins with number of pairs smaller than this value are ignored. Defaults to NULL.

bin.cloud
logical. If TRUE and option = "bin" the cloud values for each class are included in the output. Defaults to FALSE.

direction
a vector with values of 4 angles, indicating the directions for which the variograms will be computed. Default corresponds to c(0, 45, 90, 135) (degrees).

tolerance
numerical value for the tolerance angle, when computing directional variograms. The value must be in the interval [0, 90] degrees. Defaults to pi/8.

unit.angle
defines the unit for the specification of angles in the two previous arguments. Options are "degrees" and "radians".

messages
logical. Indicates whether status messages should be printed on the screen (or output device) while the function is running.

...
arguments to be passed to the function ksmooth, if option = "smooth".

Value

The output is an object of the class variog4, a list with five components. The first four elements are estimated variograms for the directions provided and the last is the omnidirectional variogram. Each individual component is an object of the class variogram, an output of the function variog.

Author(s)

Paulo J. Ribeiro Jr. paulojus@leg.ufpr.br,
Peter J. Diggle p.diggle@lancaster.ac.uk.

References

Further information on the package geoR can be found at:
http://www.leg.ufpr.br/geoR.

See Also

variog for variogram calculations and plot.variog4 for plotting results

Examples

var4 <- variog4(s100, max.dist=1)
plot(var4)

[Package geoR version 1.7-5.2.1 Index]