wrappers {geoR}	R Documentation

Wrappers for the C functions used in geoR

Description

These functions are wrappers for some (but not all) the C functions included in the geoR package.
Typically the C code is directly called from the geoR functions but these functions allows independent calls.

Usage

diffpairs(coords, data)
loccoords(coords, locations)
.diagquadraticformXAX(X, lowerA, diagA) 
.bilinearformXAY(X, lowerA, diagA, Y) 
.corr.diaglowertri(coords, cov.model, phi, kappa)
.Ccor.spatial(x, phi, kappa, cov.model)

Arguments

coords
an n x 2 matrix with the data coordinates.

data
an vector with the data values.

locations
an N x 2 matrix with the coordinates of the prediction locations.

lowerA
a vector with the diagonal terms of the symmetric matrix A.

diagA
a vector with the diagonal terms of the symmetric matrix A.

X
a matrix with conforming dimensions.

Y
a matrix with conforming dimensions.

cov.model
covariance model, see cov.spatial for options and more details.

phi
numerical value of the correlation function parameter phi.

kappa
numerical value of the correlation function parameter kappa.

x
a vector of distances.

Value

The outputs for the different functions are:

diffpairs
returns a list with elements dist - the distance between pairs of points, and diff - the difference between the values of the attributes.

loccoords
returns a n x N matrix with distances between data points and prediction locations.

diagquadraticformXAX
returns a vector with the diagonal term of the quadratic form t(X) %*% A %*% X.

bilinearformXAY
returns a vector or a matrix with the terms of the quadratic form t(X) %*% A %*% Y.

corr.diaglowertri
returns the lower triangle of the correlation matrix, including the diagonal.

Ccor.spatial
returns a vector of values of spatial correlations.

Author(s)

Paulo Justiniano Ribeiro Jr. paulojus@leg.ufpr.br,
Peter J. Diggle p.diggle@lancaster.ac.uk.

References

Further information on the package geoR can be found at:
http://www.leg.ufpr.br/geoR.

[Package geoR version 1.7-5.2.1 Index]